* Move confirm modal service to services * Move app-confirm-modal component to app-component
