+++ frontend/src/app/app.component.html
@@ -1,2 +1,5 @@
 <app-navbar></app-navbar>
 <router-outlet></router-outlet>
+
+
+<app-confirm-modal></app-confirm-modal>
