+++ frontend/src/app/services/confirm-modal/confirm-modal.service.ts
@@ -1,6 +1,6 @@
 import {Injectable, OnInit} from '@angular/core';
 import {ReplaySubject, Subject} from 'rxjs';
-import {ConfirmModalConfig} from '../ui/confirm-modal/confirm-modal.component';
+import {ConfirmModalConfig} from '../../ui/confirm-modal/confirm-modal.component';
 import {take} from 'rxjs/operators';
 
 @Injectable({
@@ -9,8 +9,10 @@ import {take} from 'rxjs/operators';
 export class ConfirmModalService {
   public clickEvent$: Subject<boolean> = new Subject();
   public showModal$: ReplaySubject<boolean> = new ReplaySubject();
+  private confirmModalConfig: ConfirmModalConfig;
 
   public confirm(config: ConfirmModalConfig): Promise<boolean> {
+    this.confirmModalConfig = config;
     this.showModal$.next(true);
     return this.clickEvent$.pipe(
       take(1),
@@ -19,8 +21,9 @@ export class ConfirmModalService {
 
   constructor() {
     this.showModal$.next(false);
-    this.clickEvent$.subscribe((v) => {
-      console.log('ValueChange', v);
-    });
+  }
+
+  get getConfig(): ConfirmModalConfig {
+    return this.confirmModalConfig;
   }
 }
