+++ frontend/src/app/ui/confirm-modal/confirm-modal.component.ts
@@ -1,7 +1,7 @@
 import {Component, Input, OnDestroy, OnInit} from '@angular/core';
 import {UiButton, UiButtonGroup} from '../ui.interface';
 import {Subject, Subscription} from 'rxjs';
-import {ConfirmModalService} from '../../admin/confirm-modal.service';
+import {ConfirmModalService} from '../../services/confirm-modal/confirm-modal.service';
 
 @Component({
   selector: 'app-confirm-modal',
@@ -9,25 +9,21 @@ import {ConfirmModalService} from '../../admin/confirm-modal.service';
   styleUrls: ['./confirm-modal.component.scss']
 })
 export class ConfirmModalComponent implements OnInit, OnDestroy {
-  @Input() title = '';
-  @Input() description = '';
-  @Input() confirmAction: () => void;
-  @Input() cancelAction: () => void;
   public showModal = false;
   private showModalSubscription: Subscription;
-
-
-  public confirmButton: UiButton = {
-    icon: 'trash',
-    title: 'Löschen'
-  };
+  public config: ConfirmModalConfig;
 
   constructor(private confirmService: ConfirmModalService) {
   }
 
   ngOnInit(): void {
+    this.setDefaultValues();
     this.showModalSubscription = this.confirmService.showModal$.subscribe(
-      val => this.showModal = val
+      val => {
+        this.showModal = val;
+        const config = this.confirmService.getConfig;
+        this.config = {...this.config, ...config};
+      }
     );
   }
 
@@ -38,16 +34,29 @@ export class ConfirmModalComponent implements OnInit, OnDestroy {
   public runConfirmAction(): void {
     this.confirmService.showModal$.next(false);
     this.confirmService.clickEvent$.next(true);
+    this.setDefaultValues();
   }
 
   public runCancelAction(): void {
     this.confirmService.showModal$.next(false);
     this.confirmService.clickEvent$.next(false);
+    this.setDefaultValues();
   }
 
+  private setDefaultValues(): void {
+    this.config = {
+      title: 'Aktion bestätigen',
+      description: '',
+      confirmText: 'Bestätigen',
+      cancelText: 'Abbrechen',
+    };
+  }
 
 }
 
-export interface ConfirmModalConfig{
+export interface ConfirmModalConfig {
   title: string;
+  description?: string;
+  confirmText?: string;
+  cancelText?: string;
 }
