+++ frontend/src/app/ui/modal/modal.component.ts
@@ -2,7 +2,7 @@ import {Component, Input, OnInit} from '@angular/core';
 import {ModalService} from '../../services/modal/modal.service';
 
 @Component({
-  selector: ' app-modal',
+  selector: 'app-modal',
   templateUrl: './modal.component.html',
   styleUrls: ['./modal.component.scss']
 })
